galeria imagenes terminada y se agregaron elementos faltantes menu
