+++ package.json
@@ -11,7 +11,8 @@
     "immutability-helper": "^2.1.2",
     "react": "15.4.2",
     "react-di": "^0.3.1",
-    "react-native": "0.41.2"
+    "react-native": "0.41.2",
+    "react-native-gallery": "0.0.17"
   },
   "devDependencies": {
     "babel-jest": "18.0.0",
